pyproject.toml
[project]
name = "deep-search-agent1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "openai==1.98.0",
    "openai-agents>=0.2.5",
    "tavily-python>=0.7.10",
]
